emphasize timingcost importance
There's really no point in correcting other issues if your timingcost is above 500... Correcting all other issues would still give you an extremely slow server with a timingcost that high. This change hides all other suggestions when the timingcost is above 500 to emphasize the importance of timingcost. It'd be nice to have an override button or something if someone wants to try and implement that. Maybe a button at the bottom to say "I know what I'm doing. Show all suggestions"? More work than it's worth but if someone wants to try it I'll accept a PR